Migrate RegexTokenizer unit tests from test/experimental/test_functional.py to test/test_transforms.py
