+++ test/experimental/test_functional.py
@@ -4,9 +4,7 @@ import unittest
 
 import torch
 import torchtext.data as data
-from torchtext._torchtext import RegexTokenizer as RegexTokenizerPybind
 from torchtext.experimental.transforms import basic_english_normalize
-from torchtext.transforms import RegexTokenizer
 
 from ..common.torchtext_test_case import TorchtextTestCase
 
@@ -101,126 +99,3 @@ class TestFunctional(TorchtextTestCase):
             torch.save(ben, save_path)
             loaded_ben = torch.load(save_path)
             self.assertEqual(loaded_ben(test_sample), ref_results)
-
-    # TODO(Nayef211): remove decorator once	https://github.com/pytorch/pytorch/issues/38207 is closed
-    @unittest.skipIf(platform.system() == "Windows", "Test is known to fail on Windows.")
-    def test_RegexTokenizer(self):
-        test_sample = "'\".<br />,()!?;:   Basic Regex Tokenization for a Line of Text   '\".<br />,()!?;:"
-        ref_results = [
-            "'",
-            ".",
-            ",",
-            "(",
-            ")",
-            "!",
-            "?",
-            "Basic",
-            "Regex",
-            "Tokenization",
-            "for",
-            "a",
-            "Line",
-            "of",
-            "Text",
-            "'",
-            ".",
-            ",",
-            "(",
-            ")",
-            "!",
-            "?",
-        ]
-        patterns_list = [
-            (r"\'", " '  "),
-            (r"\"", ""),
-            (r"\.", " . "),
-            (r"<br \/>", " "),
-            (r",", " , "),
-            (r"\(", " ( "),
-            (r"\)", " ) "),
-            (r"\!", " ! "),
-            (r"\?", " ? "),
-            (r"\;", " "),
-            (r"\:", " "),
-            (r"\s+", " "),
-        ]
-
-        patterns = [pair[0] for pair in patterns_list]
-        replacements = [pair[1] for pair in patterns_list]
-        r_tokenizer = RegexTokenizer(RegexTokenizerPybind(patterns, replacements, False))
-        eager_tokens = r_tokenizer(test_sample)
-
-        jit_r_tokenizer = torch.jit.script(r_tokenizer)
-        jit_tokens = jit_r_tokenizer(test_sample)
-
-        assert not r_tokenizer.is_jitable
-        # Call the __prepare_scriptable__() func and convert the building block to the torbhind version
-        # Not expect users to use the torchbind version on eager mode but still need a CI test here.
-        assert r_tokenizer.__prepare_scriptable__().is_jitable
-
-        self.assertEqual(eager_tokens, ref_results)
-        self.assertEqual(jit_tokens, ref_results)
-
-    def test_load_and_save(self):
-        test_sample = "'\".<br />,()!?;:   Basic Regex Tokenization for a Line of Text   '\".<br />,()!?;:"
-        ref_results = [
-            "'",
-            ".",
-            ",",
-            "(",
-            ")",
-            "!",
-            "?",
-            "Basic",
-            "Regex",
-            "Tokenization",
-            "for",
-            "a",
-            "Line",
-            "of",
-            "Text",
-            "'",
-            ".",
-            ",",
-            "(",
-            ")",
-            "!",
-            "?",
-        ]
-        patterns_list = [
-            (r"\'", " '  "),
-            (r"\"", ""),
-            (r"\.", " . "),
-            (r"<br \/>", " "),
-            (r",", " , "),
-            (r"\(", " ( "),
-            (r"\)", " ) "),
-            (r"\!", " ! "),
-            (r"\?", " ? "),
-            (r"\;", " "),
-            (r"\:", " "),
-            (r"\s+", " "),
-        ]
-
-        with self.subTest("pybind"):
-            save_path = os.path.join(self.test_dir, "regex_pybind.pt")
-
-            patterns = [pair[0] for pair in patterns_list]
-            replacements = [pair[1] for pair in patterns_list]
-            tokenizer = RegexTokenizer(RegexTokenizerPybind(patterns, replacements, False))
-            torch.save(tokenizer, save_path)
-            loaded_tokenizer = torch.load(save_path)
-            results = loaded_tokenizer(test_sample)
-            self.assertEqual(results, ref_results)
-
-        with self.subTest("torchscript"):
-            save_path = os.path.join(self.test_dir, "regex_torchscript.pt")
-            # Call the __prepare_scriptable__() func and convert the building block to the torbhind version
-            # Not expect users to use the torchbind version on eager mode but still need a CI test here.
-            patterns = [pair[0] for pair in patterns_list]
-            replacements = [pair[1] for pair in patterns_list]
-            tokenizer = RegexTokenizer(RegexTokenizerPybind(patterns, replacements, False)).__prepare_scriptable__()
-            torch.save(tokenizer, save_path)
-            loaded_tokenizer = torch.load(save_path)
-            results = loaded_tokenizer(test_sample)
-            self.assertEqual(results, ref_results)
